+++ src/app.module.ts
@@ -20,6 +20,7 @@ import compressionConfig from './config/compression.config.js';
 import thumbnailConfig from './config/thumbnail.config.js';
 import cleanupConfig from './config/cleanup.config.js';
 import heavyTasksQueueConfig from './config/heavy-tasks-queue.config.js';
+import urlUploadConfig from './config/url-upload.config.js';
 import type { AppConfig } from './config/app.config.js';
 import pkg from '../package.json' with { type: 'json' };
 
@@ -35,6 +36,7 @@ import pkg from '../package.json' with { type: 'json' };
         thumbnailConfig,
         cleanupConfig,
         heavyTasksQueueConfig,
+        urlUploadConfig,
       ],
       envFilePath: [`.env.${process.env.NODE_ENV ?? 'development'}`, '.env'],
       cache: true,